requirements.txt
aepp-sdk==2.0.0
asn1crypto==0.24.0
base58==0.2.5
blessings==1.7
bpython==0.18
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
cryptography==2.7
curtsies==0.3.0
decorator==4.4.0
Django==2.1.7
django-cors-headers==3.1.0
django-extensions==2.2.1
django-rest-auth==0.9.5
djangorestframework==3.10.1
djangorestframework-jwt==1.11.0
future==0.17.1
greenlet==0.4.15
idna==2.7
namedtupled==0.3.3
paypalrestsdk==1.13.1
pycparser==2.19
Pygments==2.3.1
PyJWT==1.7.1
PyNaCl==1.3.0
pyOpenSSL==19.0.0
Pyro4==4.76
pytz==2018.9
PyYAML==5.1
requests==2.20.0
rlp==0.6.0
semver==2.8.1
serpent==1.28
six==1.12.0
typing==3.6.6
urllib3==1.24.1
validators==0.12.1
wcwidth==0.1.7
websocket-client==0.48.0
